requirements.txt
django
djangorestframework
djangorestframework-simplejwt
djoser
django-cors-headers
python-dotenv
psycopg2-binary
supabase
